feat: add habit current streak column to index page (#150)
* ui: add streak badge components

This commit implements all UI components necessary to implement streak
badges on the index page.

Signed-off-by: Konstantin Tutsch <mail@konstantintutsch.com>

* feat: move streak composing code to seperate function

Signed-off-by: Konstantin Tutsch <mail@konstantintutsch.com>

* feat: refresh habit rows on index page when streak column is enabled

Signed-off-by: Konstantin Tutsch <mail@konstantintutsch.com>

* perf: only compose streaks for habits with period on index page

Signed-off-by: Konstantin Tutsch <mail@konstantintutsch.com>

* fix: crash when period was set and habit was never completed

This commit fixes a crash that occured, when the index streak column was
enabled, the period of a habit was set and when that same habit was
never completed before.

Signed-off-by: Konstantin Tutsch <mail@konstantintutsch.com>

* fix: crash on habit page when habit was never completed

This commit fixes a crash caused by the best streaks feature that occured when a habit had a period, was never completed
and its page was opened.

Signed-off-by: Konstantin Tutsch <mail@konstantintutsch.com>

* perf: only check if composed streaks are NoneType

Signed-off-by: Konstantin Tutsch <mail@konstantintutsch.com>

* style: compare against None

This commit changes comparing the type against NoneType to comparing the
variable against None when calculating streaks.

Signed-off-by: Konstantin Tutsch <mail@konstantintutsch.com>

---------

Signed-off-by: Konstantin Tutsch <mail@konstantintutsch.com>
Co-authored-by: Henry Zhu <daya0576@users.noreply.github.com>